EPBDS-7699 Nullsafety doesn't work with DobleValues for avg() function
--HG--
branch : 5.21.x

+++ DEV/org.openl.rules/src/org/openl/rules/helpers/RulesUtils.java
@@ -1181,7 +1181,7 @@ public class RulesUtils {
      * @param values
      * @return
      */
-    public static java.math.BigInteger median(java.math.BigInteger[] values) {
+    public static java.math.BigDecimal median(java.math.BigInteger[] values) {
         return MathUtils.median(values);
     }
